Duration type migration from String to Integer, which is the duration in minutes

+++ src/main/java/dto/SearchTourRequestDTO.java
@@ -4,5 +4,5 @@ import domain.Tag;
 
 import java.util.List;
 
-public record SearchTourRequestDTO(Long cityId, Long duration, String themeName, List<Tag> tagNames) {
+public record SearchTourRequestDTO(Long cityId, Integer duration, String themeName, List<Tag> tagNames) {
 }
